refactor: rename Duke to Mama across codebase
