+++ src/test/java/seedu/mama/MamaTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.mama;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class MamaTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
